+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-## 2.0.0 (Unreleased)
+## 2.1.0 (Unreleased)
+## 2.0.0 (December 28, 2019)
 
 - **Added:** `examples` [GH-24](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/24)
 - **Added:** `examples` [GH-28](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/28)
 
 IMPROVEMENTS:
 
+- improve readme and basic examples [GH-32](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/32)
+- improve(ecs): updated README.md and added chinese version. [GH-32](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/32)
+- improve(ecs): added submodules and examples. [GH-31](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/31)
+- add modules and improve bare-metal [GH-30](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/30)
 - improve(ecs): modified the ecs module outputs. [GH-23](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/23)
 - improve(ecs): added provider settings. [GH-22](https://github.com/terraform-alicloud-modules/terraform-alicloud-ecs-instance/pull/22)